continued work on file io

+++ src/chat/model/FileController.java
@@ -1,26 +1,85 @@
 package chat.model;
 
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.net.URL;
 import java.net.URLDecoder;
-import java.nio.charset.Charset;
-import java.nio.file.FileStore;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
+import java.util.Scanner;
 
-public class FileHandler
+import chat.controller.ChatController;
+
+public class FileController
 {
 	private String directory;
 	private String fileSeparator;
 
-	public FileHandler()
+	public FileController()
 	{
 		directory = "";
 		fileSeparator = System.getProperty("file.separator");
+	}
+
+	public static void saveFile(ChatController baseController, String fileName, String contents)
+	{
+		File saveFile;
+		if (fileName.length() > 5)
+		{
+			if (fileName.substring(fileName.length() - 4).equals(".txt"))
+			{
+				saveFile = new File(fileName);
+			} else
+			{
+				saveFile = new File(fileName + ".txt");
+			}
+		} else
+		{
+			saveFile = new File("derp derp derp.txt");
+		}
+		try
+		{
+			FileWriter saveFileWriter = new FileWriter(saveFile);
+			saveFileWriter.write(contents);
+			saveFileWriter.close();
+		} catch (IOException e)
+		{
+			baseController.handleErrors(e);
+		}
+	}
+
+	public static String readFile(ChatController baseController, String fileName)
+	{
+		String fileContenets = "";
+		File readFile;
+		if (fileName.substring(fileName.length() - 4).equals(".txt"))
+		{
+			readFile = new File(fileName);
+		} else
+		{
+			readFile = new File(fileName + ".txt");
+		}
+		try
+		{
+			Scanner fileReader = new Scanner(readFile);
+			while (fileReader.hasNextLine())
+			{
+				fileContenets += fileReader.nextLine();
+			}
+		} catch (IOException e)
+		{
+			baseController.handleErrors(e);
+
+		} catch (NullPointerException e)
+		{
+			baseController.handleErrors(e);
+		}
+
+		return fileContenets;
 	}
 
 	public ArrayList<String> getData(String name)
@@ -39,6 +98,7 @@ public class FileHandler
 		}
 		return data;
 	}
+
 	public String getRawData(String name)
 	{
 		String data = "";
@@ -47,7 +107,7 @@ public class FileHandler
 		{
 			try
 			{
-				data = new String (Files.readAllBytes(path),"UTF-8");
+				data = new String(Files.readAllBytes(path), "UTF-8");
 			} catch (IOException e)
 			{
 				e.printStackTrace();
@@ -69,8 +129,10 @@ public class FileHandler
 		}
 		return success;
 	}
+
 	/**
 	 * warning if the program crahes chance of perma lock
+	 * 
 	 * @return success sate
 	 */
 	public void lock()
@@ -86,14 +148,17 @@ public class FileHandler
 			e.printStackTrace();
 		}
 	}
+
 	public void unlock()
 	{
 		try
 		{
 			File lockFile = new File(directory + fileSeparator + "Locked.txt");
-			lockFile.delete();
-		}
-		catch(Exception e)
+			if (lockFile.exists())
+			{
+				lockFile.delete();
+			}
+		} catch (Exception e)
 		{
 			e.printStackTrace();
 		}
@@ -103,11 +168,11 @@ public class FileHandler
 	{
 		boolean success = true;
 		Path path = Paths.get(directory + fileSeparator + name);
-		if(!Files.exists(path))
+		if (!Files.exists(path))
 		{
 			try
 			{
-				if(!path.toFile().createNewFile())
+				if (!path.toFile().createNewFile())
 				{
 					System.out.println("Error Data Cannot be saved");
 					success = false;
@@ -155,7 +220,7 @@ public class FileHandler
 	private String getParentDirectory() throws Exception
 	{
 		String parentDirectory = "";
-		URL path = FileHandler.class.getProtectionDomain().getCodeSource().getLocation();
+		URL path = FileController.class.getProtectionDomain().getCodeSource().getLocation();
 		String thisPath = URLDecoder.decode(path.getFile(), "UTF-8");
 		parentDirectory = new File(thisPath).getParentFile().getPath();
 		return parentDirectory;
@@ -167,11 +232,14 @@ public class FileHandler
 		return directory;
 
 	}
-	
+
 	public boolean getIsLocked()
 	{
 		boolean isLocked = false;
-		if(Files.exists(Paths.get(directory + fileSeparator + "Locked.txt"))){isLocked = true;}
+		if (Files.exists(Paths.get(directory + fileSeparator + "Locked.txt")))
+		{
+			isLocked = true;
+		}
 		return isLocked;
 	}
